Moved study import to study folder
